requirements.txt
cryptography==2.8
tqdm==4.61.1
requests==2.28.1